refactor: move Flask app logic to main.py and clean up service.py
